feat(stats): refactor stats components to use a shared base class and models.
